+++ lib/ayadn/settings.rb
@@ -116,6 +116,7 @@ module Ayadn
           conf[:tvshow] = {hashtag: 'nowwatching'} if conf[:tvshow].nil?
           conf[:formats][:list] = {reverse: true} if conf[:formats][:list].nil?
           conf[:timeline][:compact] = false if conf[:timeline][:compact].nil?
+          conf[:timeline][:word_wrap] = false if conf[:timeline][:word_wrap].nil?
           conf[:timeline][:show_channel_oembed] = true if conf[:timeline][:show_channel_oembed].nil?
           conf[:marker] = {update_messages: true} if conf[:marker].nil?
           conf[:blacklist] = {active: true} if conf[:blacklist].nil?
@@ -196,7 +197,8 @@ module Ayadn
           show_spinner: true,
           show_debug: false,
           show_channel_oembed: true,
-          compact: false
+          compact: false,
+          word_wrap: false
         },
         marker: {
           update_messages: true